agents/daily-standup-coordinator/requirements.txt
# Core Agent & LLM Dependencies
langchain>=0.2.10,<0.3.0
langgraph>=0.1.19,<0.2.0
langchain-openai>=0.1.16,<0.2.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0

# For Google Workspace Tools
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.7.0

# For Slack Tools
slack_sdk>=3.0.0
